Create odometry.py
Created odometry node to utilise encoders to track current robot
position. Also includes a service that will allow the odometry to
be reset back to 0 which could be useful when navigation steps should be
updated.

Maths needs to be checked and obviously also needs to be tested with
encoder hardware